+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 13.7.3
 
+### New Features:
+
+- Der Würfel Dialog lässt sich jetzt überspringen, indem `skill.roll(options)` mit `askUser: false` aufgerufen wird. Special Thanks to @IceFreez3r
+
 ### Fixed
 
 - Ein Fehler wurde behoben, bei welchem das Öffnen des Rüstungs-Sheets mit einem Stapelüberlauf endete.